fix: address PR #189 review comments
- Fix countStashEntries("") returning 1 instead of 0 (add early return)
- Extract ParseDependencyScope to standalone JUnit 5 test class (was
  silently skipped inside JUnit 3 BasePlatformTestCase)
- Remove conditional assertions in ClientBinaryDetectorTest (now deterministic)
- Add missing ToolCategory.OTHER assertion for PERMISSIVE preset
- Rename misleading test codexStripsUnderscorePrefix → codexLeavesCanonicalPrefixIntact
