+++ plugin-core/src/test/java/com/github/catatafishen/agentbridge/psi/tools/project/EditProjectStructureToolTest.java
@@ -1,13 +1,7 @@
 package com.github.catatafishen.agentbridge.psi.tools.project;
 
 import com.google.gson.JsonObject;
-import com.intellij.openapi.roots.DependencyScope;
 import com.intellij.testFramework.fixtures.BasePlatformTestCase;
-import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.api.Test;
-
-import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 
 /**
  * Platform tests for {@link EditProjectStructureTool}.
@@ -219,94 +213,4 @@ public class EditProjectStructureToolTest extends BasePlatformTestCase {
             result.contains("sdk_type"));
     }
 
-    // ── parseDependencyScope (private static, tested via reflection) ─────────
-
-    /**
-     * JUnit 5 tests for {@link EditProjectStructureTool}'s private static
-     * {@code parseDependencyScope(String)} method, exercised via reflection.
-     *
-     * <p>Because the outer class is JUnit 3 ({@link BasePlatformTestCase}),
-     * these tests are best run directly (right-click the inner class in IntelliJ).
-     */
-    @Nested
-    class ParseDependencyScope {
-
-        private static final Method PARSE_DEPENDENCY_SCOPE;
-
-        static {
-            try {
-                PARSE_DEPENDENCY_SCOPE = EditProjectStructureTool.class
-                    .getDeclaredMethod("parseDependencyScope", String.class);
-                PARSE_DEPENDENCY_SCOPE.setAccessible(true);
-            } catch (NoSuchMethodException e) {
-                throw new AssertionError("Method not found — signature changed?", e);
-            }
-        }
-
-        /**
-         * Invoke the private static method, unwrapping {@link InvocationTargetException}
-         * so tests see the real exception (e.g. {@link NullPointerException}).
-         */
-        private static DependencyScope invoke(String scopeStr) throws Exception {
-            try {
-                return (DependencyScope) PARSE_DEPENDENCY_SCOPE.invoke(null, scopeStr);
-            } catch (InvocationTargetException e) {
-                Throwable cause = e.getCause();
-                if (cause instanceof RuntimeException re) throw re;
-                if (cause instanceof Error err) throw err;
-                throw e;
-            }
-        }
-
-        // ── Standard uppercase values ────────────────────────────────────────
-
-        @Test
-        void compile_uppercase() throws Exception {
-            assertEquals(DependencyScope.COMPILE, invoke("COMPILE"));
-        }
-
-        @Test
-        void test_uppercase() throws Exception {
-            assertEquals(DependencyScope.TEST, invoke("TEST"));
-        }
-
-        @Test
-        void runtime_uppercase() throws Exception {
-            assertEquals(DependencyScope.RUNTIME, invoke("RUNTIME"));
-        }
-
-        @Test
-        void provided_uppercase() throws Exception {
-            assertEquals(DependencyScope.PROVIDED, invoke("PROVIDED"));
-        }
-
-        // ── Case-insensitivity (toUpperCase) ─────────────────────────────────
-
-        @Test
-        void compile_lowercase() throws Exception {
-            assertEquals(DependencyScope.COMPILE, invoke("compile"));
-        }
-
-        @Test
-        void test_lowercase() throws Exception {
-            assertEquals(DependencyScope.TEST, invoke("test"));
-        }
-
-        // ── Edge cases ───────────────────────────────────────────────────────
-
-        @Test
-        void emptyString_returnsNull() throws Exception {
-            assertNull(invoke(""));
-        }
-
-        @Test
-        void nullInput_throwsNpe() {
-            assertThrows(NullPointerException.class, () -> invoke(null));
-        }
-
-        @Test
-        void invalidScope_returnsNull() throws Exception {
-            assertNull(invoke("INVALID_SCOPE"));
-        }
-    }
 }
